+++ CHANGES.rst
@@ -8,6 +8,9 @@ Changelog
 * Interactive: add possibility to change verbosity. E.g.:
   ``client._printer.cols = 180``.  Set to ``0`` to disable.
 
+* Consider users with restricted rights.  Since Odoo 15, most
+  users don't have access to ``ir.model``.
+
 
 2.4.6 (2025-10-31)
 ~~~~~~~~~~~~~~~~~~