+++ CHANGES.rst
@@ -11,6 +11,8 @@ Changelog
 * Consider users with restricted rights.  Since Odoo 15, most
   users don't have access to ``ir.model``.
 
+* Enhance error management for JSON-2 API with Odoo >= 19.
+
 
 2.4.6 (2025-10-31)
 ~~~~~~~~~~~~~~~~~~